Make 'auto enforce DRC' respect the layers that a via features on
The 'auto enforce DRC' feature currently behaves as if vias pass
through all layers of the PCB.  This patch makes this feature check
the layers that a via passes through.

For example, if a via passes through layers 1-3 and we are routing a
trace on layer 4, then the current implementation prohibits a trace
from being routed on layer 4 through the same coordinates as the via.
With this patch the routing is permitted.

Signed-off-by: bert <bert.timmerman@xs4all.nl>